pyproject.toml
[tool.poetry]
name = "coursework-drf"
version = "0.1.0"
description = ""
authors = ["Norilka <NorilkaRus@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
ipython = "^8.21.0"
pytils = "^0.4.1"
python-dotenv = "^1.0.1"
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.4.1"
coreapi = "^2.3.3"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
stripe = "^8.2.0"
redis = "^5.0.1"
celery = "^5.3.6"
eventlet = "^0.35.1"
django = "5.0"
django-celery-beat = "2.1.0"
asgiref = "^3.7.2"
asttokens = "^2.4.1"
cron-descriptor = "^1.4.3"
flake8 = "^7.0.0"
pytest = "^8.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
